+++ src/main/java/com/protas/movieapp/screeningroom/ScreeningRoomRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ScreeningRoomRepository extends JpaRepository<ScreeningRoom, Integer> {
+public interface ScreeningRoomRepository extends JpaRepository<ScreeningRoom, Long> {
     Optional<List<ScreeningRoom>> findByCinemaId(Long cinemaId);
     Optional<ScreeningRoom> findByScreeningRoomNumberAndCinemaId(Integer screeningRoomNumber, Long cinemaId);
     void deleteByIdAndCinemaId(Long id, Long cinemaId);
-    Optional<ScreeningRoom> findByIdAndCinemaId(Integer roomId, Long cinemaId);
+    Optional<ScreeningRoom> findByIdAndCinemaId(Long roomId, Long cinemaId);
 }
